test(#5025): isolate epilogue routing regression
